Feature: Cart page development. (#9)
* Clicking Cart in AppBar goes to Cart Page

When Cart is clicked on AppBar, it goes to CartComponent, which displays Cart Items.

* Prototype: Made CartComponent display Cart Items

* 3-Grid Layout template created.

* Developed dropdown quantity menu for Cart Product

When quantity is updated, the price is also reflected reactively.

* Test case to validate change in product price based on change in quantity of the product.

* Implementing Remove button to remove the item from the Cart

This change particularly takes the move event and then calls CartService which then handles the logic to remove the cart Item.

* Refactoring the Cart Component Code

* Implemented Cart Checkout components

such as itemsCost, tax and total cost. This is exposed via the CartService's Cart API that the react components can subscribe to.

* Test: Cart cost changes based on Product quantity

When the Product in the cart changes its quantity, we expect the total charge shown for that product to reflect the change.

Implemented a unit test case for this

* CartTotalComp developed to display total cost.

When the shopper puts lot of products in the cart, it is expected that the cart will reflect the change in the total price.

This is a design to such change. CartTotalComponent displays the change based on the data provided by the CartService.

* UI/UX: Added Cost in AppBarCart for convenience.

Mobile users would like the information handy in their small tiny device (responsive). Therefore, instead of making the user scroll all the way down to see the cart, we display the total cost on the top already which the users can see easily.