Enhancement (#11)
* Init commit: Enhancing UI

Funtionality wise, project looks pretty good. But we can make some little small tweaks to enhance the UI experience. This is an attempt to that.

* Implemented Navigation Menu Items

Previously, there was no functionality to make users aware where they are. With this new functionality, users are aware that they are in certain page, based on the highlighted menu Item
 displayed in the AppBar.

* added architecture of the project

* dockerizing