Implemented Filter Options Display next to  Products
When a user opens the site, user gets a list of all the products to chose from to put in the cart. Now, having a Filter options makes it easier for the users to even further browse the product.

This commit only displays the Filters on the left. This component needs to be thoroughly tested later on in the coming commits.