Response Product Filters: Collapsing view design
It is not ideal to show the list of the categories for the user to select in the mobile, as the device screen is small. Therefore, one way to tackle this is to build a collapsing view. Given a Filter Brands, we design an option to open/close the section.

It uses React hooks to consider the state. Also, material icons for add and substract are used as well to indicate wheher the filter list is currently collapsed or open.